build: Fix miss in last refactor
