+++ src/queenbees/core/content/operations.py
@@ -4,7 +4,7 @@ from django.contrib.auth import models as auth_models
 from django.db import transaction
 
 from queenbees.core.content import models as content_models
-from queenbees.core.utils import localtime, operations
+from queenbees.utils import localtime, operations
 
 
 class DuplicateArticleError(operations.OperationError):
