+++ src/queenbees/utils/operations/errors.py
