+++ src/tests/unit/utils/test_localtime.py
@@ -5,7 +5,7 @@ import pytest
 import time_machine
 from django.test import override_settings
 
-from queenbees.core.utils import localtime
+from queenbees.utils import localtime
 
 
 def assert_timezone_aware(d: localtime.datetime) -> None:
